@@ -33,7 +33,7 @@
     "dist/**/*"
   ],
   "peerDependencies": {
-    "react": "^16.14.0 || ^17.0.1"
+    "react": "^18.1.0"
   },
   "dependencies": {
     "@tolgee/core": "^4.6.1",
@@ -43,16 +43,16 @@
     "@rollup/plugin-typescript": "8.3.2",
     "@testing-library/dom": "^8.7.2",
     "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^12.1.2",
+    "@testing-library/react": "^13.3.0",
     "@types/jest": "^27.0.2",
     "@types/node": "^17.0.8",
-    "@types/react": "^17.0.1",
-    "@types/react-dom": "^17.0.1",
+    "@types/react": "^18.0.9",
+    "@types/react-dom": "^18.0.5",
     "concurrently": "7.2.1",
     "jest": "^27.2.4",
     "jest-fetch-mock": "^3.0.3",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "reflect-metadata": "^0.1.13",
     "regenerator-runtime": "^0.13.3",
     "rollup": "2.75.3",
