@@ -3,20 +3,17 @@
   "version": "4.6.1",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.9.0",
-    "@emotion/styled": "^11.8.1",
-    "@mui/material": "^5.6.0",
     "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^12.1.5",
+    "@testing-library/react": "^13.3.0",
     "@testing-library/user-event": "^13.3.0",
     "@tolgee/react": "^4.6.1",
     "@types/jest": "^27.5.1",
     "@types/node": "^17.0.36",
-    "@types/react": "17.0.43",
-    "@types/react-dom": "^17.0.14",
+    "@types/react": "^18.0.9",
+    "@types/react-dom": "^18.0.5",
     "concurrently": "^7.2.1",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-scripts": "5.0.1",
     "serve": "^13.0.2",
     "typescript": "^4.7.2"
