@@ -1,6 +1,10 @@
 import { LangSelector } from './LangSelector';
 
-export const Navbar: React.FC = ({ children }) => {
+type Props = {
+  children?: React.ReactNode;
+};
+
+export const Navbar: React.FC<Props> = ({ children }) => {
   return (
     <div className="navbar">
       {children}
